CLEAN architecture fix for Prediction
